refactor(scraping): Extract navigation lifecycle
